Move subsystems to folder, revert commit
