+++ src/main/java/frc/robot/subsystems/led/BlinkinLedSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.led;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.hardware.BlinkinDriver;
